requirements-core.txt
# FastAPI and web dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.5

# Async
httpx>=0.24.0
aiohttp>=3.8.4

# Scraping and parsing
beautifulsoup4>=4.12.0
lxml>=4.9.2
feedparser>=6.0.10

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
validators>=0.20.0
loguru>=0.7.0
tenacity>=8.2.2

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0